refactor: move DomainCounts to projectMetadata.ts
make projectStateServices more readable. Also, DomainCounts is sent with ProjectMetadata
